requirements.txt
# Telegram
python-telegram-bot>=21.6

# YouTube downloading
yt-dlp>=2024.11.18

# Audio metadata tagging
mutagen>=1.47.0

# Fuzzy search (typo tolerance)
rapidfuzz>=3.10.0

# HTTP client (for cover art fetching)
httpx>=0.27.0

# Environment variables
python-dotenv>=1.0.0

# Logging extras
colorlog>=6.9.0
